pyproject.toml
[project]
name = "strata"
version = "0.1.0"
description = "Strata API — FastAPI backend for the portfolio backtester."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "numpy>=2.4.4",
    "pandas>=2.2,<3.0",
    "psycopg-pool>=3.3.1",
    "psycopg[binary]>=3.3.4",
    "pydantic>=2.13.4",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "yfinance>=1.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "uvicorn>=0.46.0",
]

[tool.pytest.ini_options]
testpaths = ["api/tests"]
pythonpath = ["."]
